+++ packages/openapi-ts-tests/tanstack-query/v5/test/plugins.test.ts
@@ -201,6 +201,7 @@ for (const version of versions) {
           plugins: [
             {
               getQueryData: true,
+              mutationKeys: true,
               name: '@tanstack/react-query',
               setQueryData: true,
               useGetQueryData: true,