+++ packages/openapi-ts/src/plugins/zod/v3/toAst/union.ts
@@ -5,11 +5,6 @@ import type { Chain } from '../../shared/chain';
 import { tryBuildDiscriminatedUnion } from '../../shared/discriminated-union';
 import type { ZodResult } from '../../shared/types';
 
-type UnionToAstOptions = Pick<
-  UnionResolverContext,
-  'childResults' | 'parentSchema' | 'plugin' | 'schemas'
->;
-
 function baseNode(ctx: UnionResolverContext): Chain {
   const { childResults, parentSchema, plugin, schemas, symbols } = ctx;
   const { z } = symbols;
@@ -31,7 +26,7 @@ function baseNode(ctx: UnionResolverContext): Chain {
   }
 
   if (nonNullItems.length === 1) {
-    return nonNullItems[0]!.expression;
+    return nonNullItems[0]!.chain;
   }
 
   const discriminatedExpression = tryBuildDiscriminatedUnion({
@@ -68,7 +63,7 @@ function baseNode(ctx: UnionResolverContext): Chain {
     .call(
       $.array()
         .pretty()
-        .elements(...nonNullItems.map((item) => item.expression)),
+        .elements(...nonNullItems.map((item) => item.chain)),
     );
 }
 
@@ -78,9 +73,15 @@ function unionResolver(ctx: UnionResolverContext): Chain {
   return ctx.chain.current;
 }
 
-export function unionToAst({ childResults, parentSchema, plugin, schemas }: UnionToAstOptions): {
+export function unionToAst({
+  childResults,
+  parentSchema,
+  path,
+  plugin,
+  schemas,
+}: Pick<UnionResolverContext, 'childResults' | 'parentSchema' | 'path' | 'plugin' | 'schemas'>): {
+  chain: Chain;
   childResults: Array<ZodResult>;
-  expression: Chain;
 } {
   const z = plugin.external('zod.z');
 
@@ -94,6 +95,7 @@ export function unionToAst({ childResults, parentSchema, plugin, schemas }: Unio
       base: baseNode,
     },
     parentSchema,
+    path,
     plugin,
     schema: parentSchema,
     schemas,
@@ -103,10 +105,10 @@ export function unionToAst({ childResults, parentSchema, plugin, schemas }: Unio
   };
 
   const resolver = plugin.config['~resolvers']?.union;
-  const expression = resolver?.(ctx) ?? unionResolver(ctx);
+  const chain = resolver?.(ctx) ?? unionResolver(ctx);
 
   return {
+    chain,
     childResults,
-    expression,
   };
 }